fix  -32bit and 64bit  are not  compatible  bug

+++ itcrm/MaintFormViewController.m
@@ -355,10 +355,10 @@
     return nil;
 }
 
--(float)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath{
+-(CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath{
     return 40;
 }
--(float)tableView:(SKSTableView *)tableView heightForSubRowAtIndexPath:(NSIndexPath *)indexPath{
+-(CGFloat)tableView:(SKSTableView *)tableView heightForSubRowAtIndexPath:(NSIndexPath *)indexPath{
     CGFloat height;
     //提取每行的数据
     NSMutableDictionary *dic=alist_filtered_data[indexPath.section][indexPath.subRow-1];
